Update piemain.py
The original pieparty script was modified to make it callable as a function. See setup.py and /conda.recipe/meta.yaml; the line consol scripts line enables the script to be called as a function from a module. Pieparty = pieparty.scripts.piemain:main. Where pieparty.scripts.piemain is the module and main is the function.